refactor(notifications): decompose NotificationsService into Transport/Dispatch/Reminder/Stats services
Behavior-preserving (0 $tx). Facade keeps the 15-method surface + 7-arg ctor +
delegates; internally constructs the 4 sub-services (keeps all 12 specs + module +
16 external consumers untouched). send() IN_APP+compliance gate ordering preserved.
